app_atlas/requirements.txt
asgiref==3.5.2
astroid==2.11.6
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
dill==0.3.5.1
Django==4.0.4
django-amplitude==0.7.1
django-cors-headers==3.13.0
django-etc==1.3.1
django-filter==21.1
django-hitcount==1.3.5
djangorestframework==3.13.1
flake8==4.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.7
mccabe==0.6.1
packaging==21.3
Pillow==9.1.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.14.4
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
ua-parser==0.15.0
urllib3==1.26.9
user-agents==2.2.0
wrapt==1.14.1